remove crude patch and do it the right way
